backend/requirements.txt
fastapi>=0.135.0,<1.0.0
uvicorn[standard]>=0.41.0,<1.0.0
python-dotenv>=1.2.2,<2.0.0
PyGithub>=2.8.0,<3.0.0
requests>=2.32.5,<3.0.0
